requirements.txt
PyOpenGL~=3.1.5
glfw~=2.0.0
numpy==1.19.3
scipy~=1.5.4
coloredlogs~=14.0